area_backend/requirements.txt
flasgger==0.9.5
Flask==2.2.2
Flask_Cors==3.0.10
Flask_PyMongo==2.3.0
google_api_python_client==2.77.0
google_auth_oauthlib==0.7.1
protobuf==4.21.12
PyJWT==2.6.0
pymongo==4.3.2
requests==2.25.1
flask-swagger-ui
datetime
O365
msal
msgraph-beta-sdk
pycryptodome
requests.auth
azure-ad-verify-token
gunicorn
flask_caching
slack_sdk
aiohttp
